pyproject.toml
[project]
name = "vectordb-migration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "tqdm>=4.67.1",
]
